#63: Gutted utilities & made it TypeScript
