+++ scoutradioz-utilities/package.json
@@ -1,8 +1,8 @@
 {
   "name": "@firstteam102/scoutradioz-utilities",
-  "version": "1.1.7",
+  "version": "1.1.7-test1",
   "description": "Database utilities package for Scoutradioz",
-  "main": "utilities.js",
+  "main": "build/utilities.js",
   "scripts": {
     "test": "node test.js"
   },
@@ -17,17 +17,21 @@
   },
   "homepage": "https://github.com/firstteam102/ScoringApp-Serverless#readme",
   "devDependencies": {
-    "log4js": "^6.1.2"
+    "log4js": "^6.1.2",
+	"mongodb": "^4.5.0"
   },
   "peerDependencies": {
     "mongodb": "~4.1.0"
   },
   "files": [
-    "utilities.js"
+    "src",
+	"build"
   ],
   "dependencies": {
     "@log4js-node/log4js-api": "^1.0.2",
-    "node-rest-client": "3.1.1",
-    "node-cache": "~5.1.2"
+    "@types/express": "^4.17.13",
+	"@types/node": "^16.x.x",
+    "node-cache": "~5.1.2",
+    "node-rest-client": "3.1.1"
   }
 }
