+++ scoutradioz-utilities/test.js
@@ -1,5 +1,5 @@
 const log4js = require('log4js');
-const utilities = require('./utilities');
+const utilities = require('./build/utilities');
 
 //log4js config
 log4js.configure({
@@ -27,7 +27,7 @@ logger.level = 'trace';
 utilities.config({
 	app: {url: 'mongodb://localhost:27017/app'},
 	dev: {url: 'mongodb://localhost:27017/dev'}
-});
+}, {cache: {enable: true}, debug: true});
 
 //Test utilities.getURL
 async function testGetUrl() {
@@ -72,18 +72,18 @@ async function testMultipleDbs(){
 	logger.info(`writeResult=${JSON.stringify(writeResult)}`);
 	
 	var obj = await utilities.findOne('test', {'foo': 'bar'});
-	logger.info(`In dev: ${obj}`);
+	logger.info(`In dev: ${JSON.stringify(obj)}`);
 	
 	process.env.TIER = 'app';
 	utilities.refreshTier();
 	obj = await utilities.findOne('test', {'foo': 'bar'});
-	logger.info(`In app: ${obj}`);
+	logger.info(`In app: ${JSON.stringify(obj)}`);
 	
 	process.env.TIER = 'dev';
 	utilities.refreshTier();
 	
 	var delResult = await utilities.remove('test', {'foo': 'bar'});
-	logger.info(`delResult: ${delResult}`);
+	logger.info(`delResult: ${JSON.stringify(delResult)}`);
 	
 	logger.info('Done');
 }
